+++ Modules/Intent.Modules.VisualStudio.Projects/Api/Extensions/AzureFunctionsProjectModelStereotypeExtensions.cs
@@ -14,14 +14,14 @@ namespace Intent.Modules.VisualStudio.Projects.Api
     {
         public static NETCoreSettings GetNETCoreSettings(this AzureFunctionsProjectModel model)
         {
-            var stereotype = model.GetStereotype("5d85edfb-490c-42c2-81ef-c57c0300eb5a");
+            var stereotype = model.GetStereotype(NETCoreSettings.DefinitionId);
             return stereotype != null ? new NETCoreSettings(stereotype) : null;
         }
 
 
         public static bool HasNETCoreSettings(this AzureFunctionsProjectModel model)
         {
-            return model.HasStereotype("5d85edfb-490c-42c2-81ef-c57c0300eb5a");
+            return model.HasStereotype(NETCoreSettings.DefinitionId);
         }
 
         public static bool TryGetNETCoreSettings(this AzureFunctionsProjectModel model, out NETCoreSettings stereotype)
@@ -32,20 +32,20 @@ namespace Intent.Modules.VisualStudio.Projects.Api
                 return false;
             }
 
-            stereotype = new NETCoreSettings(model.GetStereotype("5d85edfb-490c-42c2-81ef-c57c0300eb5a"));
+            stereotype = new NETCoreSettings(model.GetStereotype(NETCoreSettings.DefinitionId));
             return true;
         }
 
         public static AzureFunctionsProjectSettings GetAzureFunctionsProjectSettings(this AzureFunctionsProjectModel model)
         {
-            var stereotype = model.GetStereotype("97218a15-c5ad-42f7-aac4-a9b8edcb5acd");
+            var stereotype = model.GetStereotype(AzureFunctionsProjectSettings.DefinitionId);
             return stereotype != null ? new AzureFunctionsProjectSettings(stereotype) : null;
         }
 
 
         public static bool HasAzureFunctionsProjectSettings(this AzureFunctionsProjectModel model)
         {
-            return model.HasStereotype("97218a15-c5ad-42f7-aac4-a9b8edcb5acd");
+            return model.HasStereotype(AzureFunctionsProjectSettings.DefinitionId);
         }
 
         public static bool TryGetAzureFunctionsProjectSettings(this AzureFunctionsProjectModel model, out AzureFunctionsProjectSettings stereotype)
@@ -56,20 +56,20 @@ namespace Intent.Modules.VisualStudio.Projects.Api
                 return false;
             }
 
-            stereotype = new AzureFunctionsProjectSettings(model.GetStereotype("97218a15-c5ad-42f7-aac4-a9b8edcb5acd"));
+            stereotype = new AzureFunctionsProjectSettings(model.GetStereotype(AzureFunctionsProjectSettings.DefinitionId));
             return true;
         }
 
         public static CSharpProjectOptions GetCSharpProjectOptions(this AzureFunctionsProjectModel model)
         {
-            var stereotype = model.GetStereotype("d9b43864-0370-401b-9e71-b1d286919500");
+            var stereotype = model.GetStereotype(CSharpProjectOptions.DefinitionId);
             return stereotype != null ? new CSharpProjectOptions(stereotype) : null;
         }
 
 
         public static bool HasCSharpProjectOptions(this AzureFunctionsProjectModel model)
         {
-            return model.HasStereotype("d9b43864-0370-401b-9e71-b1d286919500");
+            return model.HasStereotype(CSharpProjectOptions.DefinitionId);
         }
 
         public static bool TryGetCSharpProjectOptions(this AzureFunctionsProjectModel model, out CSharpProjectOptions stereotype)
@@ -80,13 +80,14 @@ namespace Intent.Modules.VisualStudio.Projects.Api
                 return false;
             }
 
-            stereotype = new CSharpProjectOptions(model.GetStereotype("d9b43864-0370-401b-9e71-b1d286919500"));
+            stereotype = new CSharpProjectOptions(model.GetStereotype(CSharpProjectOptions.DefinitionId));
             return true;
         }
 
         public class NETCoreSettings
         {
             private IStereotype _stereotype;
+            public const string DefinitionId = "5d85edfb-490c-42c2-81ef-c57c0300eb5a";
 
             public NETCoreSettings(IStereotype stereotype)
             {
@@ -287,6 +288,7 @@ namespace Intent.Modules.VisualStudio.Projects.Api
         public class AzureFunctionsProjectSettings
         {
             private IStereotype _stereotype;
+            public const string DefinitionId = "97218a15-c5ad-42f7-aac4-a9b8edcb5acd";
 
             public AzureFunctionsProjectSettings(IStereotype stereotype)
             {
@@ -349,6 +351,7 @@ namespace Intent.Modules.VisualStudio.Projects.Api
         public class CSharpProjectOptions
         {
             private IStereotype _stereotype;
+            public const string DefinitionId = "d9b43864-0370-401b-9e71-b1d286919500";
 
             public CSharpProjectOptions(IStereotype stereotype)
             {