+++ Modules/Intent.Modules.VisualStudio.Projects/Api/Extensions/CSharpProjectNETModelStereotypeExtensions.cs
@@ -14,14 +14,14 @@ namespace Intent.Modules.VisualStudio.Projects.Api
     {
         public static NETSettings GetNETSettings(this CSharpProjectNETModel model)
         {
-            var stereotype = model.GetStereotype("a490a23f-5397-40a1-a3cb-6da7e0b467c0");
+            var stereotype = model.GetStereotype(NETSettings.DefinitionId);
             return stereotype != null ? new NETSettings(stereotype) : null;
         }
 
 
         public static bool HasNETSettings(this CSharpProjectNETModel model)
         {
-            return model.HasStereotype("a490a23f-5397-40a1-a3cb-6da7e0b467c0");
+            return model.HasStereotype(NETSettings.DefinitionId);
         }
 
         public static bool TryGetNETSettings(this CSharpProjectNETModel model, out NETSettings stereotype)
@@ -32,20 +32,20 @@ namespace Intent.Modules.VisualStudio.Projects.Api
                 return false;
             }
 
-            stereotype = new NETSettings(model.GetStereotype("a490a23f-5397-40a1-a3cb-6da7e0b467c0"));
+            stereotype = new NETSettings(model.GetStereotype(NETSettings.DefinitionId));
             return true;
         }
 
         public static CSharpProjectOptions GetCSharpProjectOptions(this CSharpProjectNETModel model)
         {
-            var stereotype = model.GetStereotype("d9b43864-0370-401b-9e71-b1d286919500");
+            var stereotype = model.GetStereotype(CSharpProjectOptions.DefinitionId);
             return stereotype != null ? new CSharpProjectOptions(stereotype) : null;
         }
 
 
         public static bool HasCSharpProjectOptions(this CSharpProjectNETModel model)
         {
-            return model.HasStereotype("d9b43864-0370-401b-9e71-b1d286919500");
+            return model.HasStereotype(CSharpProjectOptions.DefinitionId);
         }
 
         public static bool TryGetCSharpProjectOptions(this CSharpProjectNETModel model, out CSharpProjectOptions stereotype)
@@ -56,13 +56,14 @@ namespace Intent.Modules.VisualStudio.Projects.Api
                 return false;
             }
 
-            stereotype = new CSharpProjectOptions(model.GetStereotype("d9b43864-0370-401b-9e71-b1d286919500"));
+            stereotype = new CSharpProjectOptions(model.GetStereotype(CSharpProjectOptions.DefinitionId));
             return true;
         }
 
         public class NETSettings
         {
             private IStereotype _stereotype;
+            public const string DefinitionId = "a490a23f-5397-40a1-a3cb-6da7e0b467c0";
 
             public NETSettings(IStereotype stereotype)
             {
@@ -495,6 +496,7 @@ namespace Intent.Modules.VisualStudio.Projects.Api
         public class CSharpProjectOptions
         {
             private IStereotype _stereotype;
+            public const string DefinitionId = "d9b43864-0370-401b-9e71-b1d286919500";
 
             public CSharpProjectOptions(IStereotype stereotype)
             {