+++ Modules/Intent.Modules.VisualStudio.Projects/Api/Extensions/ClassLibraryNETCoreModelStereotypeExtensions.cs
@@ -14,14 +14,14 @@ namespace Intent.Modules.VisualStudio.Projects.Api
     {
         public static NETCoreSettings GetNETCoreSettings(this ClassLibraryNETCoreModel model)
         {
-            var stereotype = model.GetStereotype("5d85edfb-490c-42c2-81ef-c57c0300eb5a");
+            var stereotype = model.GetStereotype(NETCoreSettings.DefinitionId);
             return stereotype != null ? new NETCoreSettings(stereotype) : null;
         }
 
 
         public static bool HasNETCoreSettings(this ClassLibraryNETCoreModel model)
         {
-            return model.HasStereotype("5d85edfb-490c-42c2-81ef-c57c0300eb5a");
+            return model.HasStereotype(NETCoreSettings.DefinitionId);
         }
 
         public static bool TryGetNETCoreSettings(this ClassLibraryNETCoreModel model, out NETCoreSettings stereotype)
@@ -32,20 +32,20 @@ namespace Intent.Modules.VisualStudio.Projects.Api
                 return false;
             }
 
-            stereotype = new NETCoreSettings(model.GetStereotype("5d85edfb-490c-42c2-81ef-c57c0300eb5a"));
+            stereotype = new NETCoreSettings(model.GetStereotype(NETCoreSettings.DefinitionId));
             return true;
         }
 
         public static CSharpProjectOptions GetCSharpProjectOptions(this ClassLibraryNETCoreModel model)
         {
-            var stereotype = model.GetStereotype("d9b43864-0370-401b-9e71-b1d286919500");
+            var stereotype = model.GetStereotype(CSharpProjectOptions.DefinitionId);
             return stereotype != null ? new CSharpProjectOptions(stereotype) : null;
         }
 
 
         public static bool HasCSharpProjectOptions(this ClassLibraryNETCoreModel model)
         {
-            return model.HasStereotype("d9b43864-0370-401b-9e71-b1d286919500");
+            return model.HasStereotype(CSharpProjectOptions.DefinitionId);
         }
 
         public static bool TryGetCSharpProjectOptions(this ClassLibraryNETCoreModel model, out CSharpProjectOptions stereotype)
@@ -56,13 +56,14 @@ namespace Intent.Modules.VisualStudio.Projects.Api
                 return false;
             }
 
-            stereotype = new CSharpProjectOptions(model.GetStereotype("d9b43864-0370-401b-9e71-b1d286919500"));
+            stereotype = new CSharpProjectOptions(model.GetStereotype(CSharpProjectOptions.DefinitionId));
             return true;
         }
 
         public class NETCoreSettings
         {
             private IStereotype _stereotype;
+            public const string DefinitionId = "5d85edfb-490c-42c2-81ef-c57c0300eb5a";
 
             public NETCoreSettings(IStereotype stereotype)
             {
@@ -263,6 +264,7 @@ namespace Intent.Modules.VisualStudio.Projects.Api
         public class CSharpProjectOptions
         {
             private IStereotype _stereotype;
+            public const string DefinitionId = "d9b43864-0370-401b-9e71-b1d286919500";
 
             public CSharpProjectOptions(IStereotype stereotype)
             {