+++ Modules/Intent.Modules.VisualStudio.Projects/Api/Extensions/VisualStudioSolutionModelStereotypeExtensions.cs
@@ -14,14 +14,14 @@ namespace Intent.Modules.VisualStudio.Projects.Api
     {
         public static VisualStudioSolutionOptions GetVisualStudioSolutionOptions(this VisualStudioSolutionModel model)
         {
-            var stereotype = model.GetStereotype("90085088-b29b-4386-b87e-f7f62c76c5de");
+            var stereotype = model.GetStereotype(VisualStudioSolutionOptions.DefinitionId);
             return stereotype != null ? new VisualStudioSolutionOptions(stereotype) : null;
         }
 
 
         public static bool HasVisualStudioSolutionOptions(this VisualStudioSolutionModel model)
         {
-            return model.HasStereotype("90085088-b29b-4386-b87e-f7f62c76c5de");
+            return model.HasStereotype(VisualStudioSolutionOptions.DefinitionId);
         }
 
         public static bool TryGetVisualStudioSolutionOptions(this VisualStudioSolutionModel model, out VisualStudioSolutionOptions stereotype)
@@ -32,13 +32,14 @@ namespace Intent.Modules.VisualStudio.Projects.Api
                 return false;
             }
 
-            stereotype = new VisualStudioSolutionOptions(model.GetStereotype("90085088-b29b-4386-b87e-f7f62c76c5de"));
+            stereotype = new VisualStudioSolutionOptions(model.GetStereotype(VisualStudioSolutionOptions.DefinitionId));
             return true;
         }
 
         public class VisualStudioSolutionOptions
         {
             private IStereotype _stereotype;
+            public const string DefinitionId = "90085088-b29b-4386-b87e-f7f62c76c5de";
 
             public VisualStudioSolutionOptions(IStereotype stereotype)
             {