+++ Modules/Intent.Modules.VisualStudio.Projects/Api/ImplicitUsingModel.cs
@@ -1,0 +1,78 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using Intent.Metadata.Models;
+using Intent.Modules.Common;
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.ModuleBuilder.Templates.Api.ApiElementModel", Version = "1.0")]
+
+namespace Intent.Modules.VisualStudio.Projects.Api
+{
+    [IntentManaged(Mode.Fully, Signature = Mode.Fully)]
+    public class ImplicitUsingModel : IMetadataModel, IHasStereotypes, IHasName, IElementWrapper
+    {
+        public const string SpecializationType = "Implicit Using";
+        public const string SpecializationTypeId = "7fe449d0-5448-4ee0-9c3c-2c1226d04c5f";
+        protected readonly IElement _element;
+
+        [IntentManaged(Mode.Fully)]
+        public ImplicitUsingModel(IElement element, string requiredType = SpecializationType)
+        {
+            if (!requiredType.Equals(element.SpecializationType, StringComparison.InvariantCultureIgnoreCase))
+            {
+                throw new Exception($"Cannot create a '{GetType().Name}' from element with specialization type '{element.SpecializationType}'. Must be of type '{SpecializationType}'");
+            }
+            _element = element;
+        }
+
+        public string Id => _element.Id;
+
+        public string Name => _element.Name;
+
+        public string Comment => _element.Comment;
+
+        public IEnumerable<IStereotype> Stereotypes => _element.Stereotypes;
+
+        public IElement InternalElement => _element;
+
+        public override string ToString()
+        {
+            return _element.ToString();
+        }
+
+        public bool Equals(ImplicitUsingModel other)
+        {
+            return Equals(_element, other?._element);
+        }
+
+        public override bool Equals(object obj)
+        {
+            if (ReferenceEquals(null, obj)) return false;
+            if (ReferenceEquals(this, obj)) return true;
+            if (obj.GetType() != this.GetType()) return false;
+            return Equals((ImplicitUsingModel)obj);
+        }
+
+        public override int GetHashCode()
+        {
+            return (_element != null ? _element.GetHashCode() : 0);
+        }
+    }
+
+    [IntentManaged(Mode.Fully)]
+    public static class ImplicitUsingModelExtensions
+    {
+
+        public static bool IsImplicitUsingModel(this ICanBeReferencedType type)
+        {
+            return type != null && type is IElement element && element.SpecializationTypeId == ImplicitUsingModel.SpecializationTypeId;
+        }
+
+        public static ImplicitUsingModel AsImplicitUsingModel(this ICanBeReferencedType type)
+        {
+            return type.IsImplicitUsingModel() ? new ImplicitUsingModel((IElement)type) : null;
+        }
+    }
+}