+++ Modules/Intent.Modules.VisualStudio.Projects/ApiMetadataDesignerExtensions.cs
@@ -9,6 +9,7 @@ namespace Intent.Modules.VisualStudio.Projects.Api
 {
     public static class ApiMetadataDesignerExtensions
     {
+        public const string VisualStudioDesignerId = "0701433c-36c0-4569-b1f4-9204986b587d";
         public static IDesigner VisualStudio(this IMetadataManager metadataManager, IApplication application)
         {
             return metadataManager.VisualStudio(application.Id);
@@ -16,7 +17,7 @@ namespace Intent.Modules.VisualStudio.Projects.Api
 
         public static IDesigner VisualStudio(this IMetadataManager metadataManager, string applicationId)
         {
-            return metadataManager.GetDesigner(applicationId, "Visual Studio");
+            return metadataManager.GetDesigner(applicationId, VisualStudioDesignerId);
         }
 
     }